chore: update generic service doc, unify utils in express
